Sep 23 2022 revision requests to DSS2022
