+++ etc/dss2022/convert.py
@@ -1,21 +1,34 @@
 # this version of the 'convert' script for DSS 2022
-# expects to import array columns as postgres arrays.
+# expects to concatenate a few columns in the input
+# thereby reduce the number of columns in the output
 # see create-nextview.sql
 
 import csv
 import sys
 import re
 
-columns = [8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 24, 25]
-# [3,4,5,6,11,12,13,14,15,16,17,18,19,20,22,23,25]
-
-def encapsulate(cell):
-    lst = cell.split('|')
-    lst = ["'" + re.sub(r"('\]\[\}\{)", "\\\1", l) + "'" for l in lst]
-    arr = '{ ' + ','.join(lst) + ' }'
-    if arr == "{ '' }" : arr = ''
-    return arr
-
+concatenate_cols = [
+    [0],
+    [1],
+    [2],
+    [3, 4],
+    [6],
+    [8],
+    [10, 20],         # person
+    [11, 5, 15, 19],  # place
+    [13, 14],         # maker = org + person
+    [9, 21],          # org
+    [25],
+    [12],
+    [15],
+    [16],
+    [17],
+    [18],
+    [22],
+    [23],
+    [24],
+    [7]
+]
 
 txt_file = sys.argv[1]
 csv_file = sys.argv[2]
@@ -25,8 +38,7 @@ with open(txt_file, "r") as in_text:
     with open(csv_file, "w") as out_csv:
         out_writer = csv.writer(out_csv, escapechar='\\', quoting=csv.QUOTE_NONE, delimiter='\t')
         for row in in_reader:
-            for i, cell in enumerate(row):
-                if i in columns:
-                    row[i] = encapsulate(cell)
-            out_writer.writerow(row)
-
+            output_row = []
+            for cols in concatenate_cols:
+                output_row.append('|'.join([row[c] for c in cols]))
+            out_writer.writerow(output_row)
