+++ etc/dss2022/copy-v1.sql
